[codeowners] Fixes, works again
